(apimocker:refactor) more build fixes 😭
